requirements.txt
# Google Maps Scraper - CLI Version Requirements
# Install with: pip install -r requirements.txt

# Core scraping dependencies
selenium>=4.0.0
webdriver-manager>=3.8.0

# Additional utilities
requests>=2.25.0
urllib3>=1.26.0

# HTML parsing and data handling
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Reliability, config, typing
tenacity>=8.2.0
pyyaml>=6.0.0
pydantic>=2.0.0

# User-Agent management (optional; we also ship static list)
fake-useragent>=1.4.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
responses>=0.24.0

# Dashboard UI (optional)
streamlit>=1.28.0
pandas>=2.0.0

# Backend API (v3.0)
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Phone extraction (v3.0)
phonenumbers>=8.13.0
pytesseract>=0.3.10

# Image processing for OCR
Pillow>=10.0
opencv-python>=4.8.0

# Push notifications
pywebpush>=1.14.0
py-vapid>=1.9.0
stripe>=7.0.0

# Screenshot capture
mss>=9.0

# System monitoring
psutil>=5.9.0

# E2E Testing
playwright>=1.40.0
pytest-playwright>=0.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Export formats
openpyxl>=3.1.0

# Performance & Database
psycopg2-binary>=2.9.0
brotli>=1.1.0
sqlalchemy>=2.0.0
prometheus-client>=0.19.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
websocket-client>=1.6.0

# Enrichment & Verification (Phase 3)
twilio>=8.10.0  # For phone verification
openai>=1.0.0  # For AI enhancements (optional)
anthropic>=0.3.0  # For Claude AI (alternative to OpenAI)

# Performance & Async (Phase 4)
httpx>=0.25.0  # For async HTTP requests
aiohttp>=3.9.0  # Alternative async HTTP client


